requirements.txt
pyyaml
paramiko
requests
colorlog